move basic command stuff to abstract class
